added the fix for the comments
