+++ modules/.submodules.json
@@ -4,6 +4,7 @@
       "33acrossIdSystem",
       "admixerIdSystem",
       "adtelligentIdSystem",
+      "adqueryIdSystem",
       "amxIdSystem",
       "britepoolIdSystem",
       "connectIdSystem",
@@ -13,9 +14,11 @@
       "deepintentDpesIdSystem",
       "dmdIdSystem",
       "fabrickIdSystem",
+      "freepassIdSystem",
       "hadronIdSystem",
       "id5IdSystem",
       "ftrackIdSystem",
+      "gravitoIdSystem",
       "identityLinkIdSystem",
       "idxIdSystem",
       "imuIdSystem",
@@ -23,13 +26,16 @@
       "justIdSystem",
       "kinessoIdSystem",
       "liveIntentIdSystem",
+      "lockrAIMIdSystem",
       "lotamePanoramaIdSystem",
       "merkleIdSystem",
       "mwOpenLinkIdSystem",
+      "mygaruIdSystem",
       "naveggIdSystem",
       "netIdSystem",
       "novatiqIdSystem",
       "oneKeyIdSystem",
+      "operaadsIdSystem",
       "parrableIdSystem",
       "pubProvidedIdSystem",
       "publinkIdSystem",
@@ -43,13 +49,7 @@
       "euidIdSystem",
       "unifiedIdSystem",
       "verizonMediaIdSystem",
-      "zeotapIdPlusIdSystem",
-      "adqueryIdSystem",
-      "gravitoIdSystem",
-      "freepassIdSystem",
-      "operaadsIdSystem",
-      "mygaruIdSystem",
-      "lockrAIMIdSystem"
+      "zeotapIdPlusIdSystem"
     ],
     "adpod": [
       "freeWheelAdserverVideo",
